+++ client/coresource/src/ru/spbstu/gyboml/core/destructible/DestructionListener.java
@@ -1,5 +1,5 @@
 package ru.spbstu.gyboml.core.destructible;
 
 public interface DestructionListener {
-    void destructionOccured(int newHP);
+    void destructionOccured(float newHP);
 }
